backend/requirements.txt
google-generativeai==0.8.3
google-genai==1.41.0
pillow==10.1.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
playwright==1.40.0
google-search-results==2.4.2